requirements.txt
python-telegram-bot>=22.0,<23.0
openai>=1.76.0,<2.0.0
python-dotenv
aiofiles>=24.0.0